AArch64: Eliminate redundant constant load instructions.
This patch fixes issue: https://github.com/oracle/graal/issues/2216

Unused constant load won't be deleted after match rules. As a result,
it might generate redundant "mov/orr" instructions if the constants
cannot be inlined to codes.

For example, the original codes:

  lsr    w0, w0, #6
  orr    w1, wzr, #0xfffff
  and    w0, w0, w1

will be optimized to the following pattern by adding match rules:

  orr    w1, wzr, #0xfffff
  ubfx   w0, w0, #6, #20

However, the constant "0xfffff" is not deleted even though it will
never be used. To fix it, this patch inserts a phase applied to
LIRInstructions. It's used to eliminate the unused constant load ops
before code generation. Although it is limited to constant load ops
currently, it can also be extended to remove other kinds of unused
instructions if needed in future.

Change-Id: Ibed66714c12e02f400ab494683dee8826ede01a1